requirements.txt
torch
transformers
faiss-cpu
Pillow
matplotlib
